backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
requests
python-dotenv
sqlalchemy>=2.0
psycopg2-binary>=2.9
alembic>=1.13
